% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{addDistance}
\alias{addDistance}
\title{addDistance}
\usage{
addDistance(ranges, maxDistance)
}
\arguments{
\item{ranges}{A GRange object which were created during
identifyClusters function}

\item{maxDistance}{A number with the maximum distance between DNA mutations
that are defined as being in a cluster.}
}
\value{
A GRange object with added distance and a logical column as metadata
}
\description{
A function that adds distances to nearest mutation to the tibble
  of the identifyClusters function
}
