% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyAndAnnotateClusters.R
\name{addIdentTableComment}
\alias{addIdentTableComment}
\title{addIdentTableComment}
\usage{
addIdentTableComment(dataTable, linkPatterns)
}
\description{
Adds commentary to the results of the
  \code{\link{identifyAndAnnotateClusters}} function.
}
