% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{addShuffleMutationComment}
\alias{addShuffleMutationComment}
\title{addShuffleMutationComment}
\usage{
addShuffleMutationComment(results, searchIdHeader)
}
\arguments{
\item{results}{The results from the shuffleMutation function}

\item{searchIdHeader}{Name of the the column with mutation processes}
}
\description{
A function to add comment about the result table.
}
