% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{addToResultTable}
\alias{addToResultTable}
\title{addToResultTable}
\usage{
addToResultTable(clusterTable, searchPatterns, searchIdHeader)
}
\arguments{
\item{clusterTable}{The outcome of the groupClusters functions}

\item{searchPatterns}{A tibble with the known mutation patterns. The
data/mutationPatterns.rds is the default search table.}
}
\description{
A function to create a table with frequencies of patterns.
}
