% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{callLinkPatterns}
\alias{callLinkPatterns}
\title{callLinkPatterns}
\usage{
callLinkPatterns(x, linkedVariables)
}
\arguments{
\item{x}{contains mutation information}

\item{linkedVariables}{Contains all the parameters needed voor
\code{\link{linkPatterns}}}
}
\description{
A function to correctly call the linkPattern function
}
