% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{checkLinkPatternsParameters}
\alias{checkLinkPatternsParameters}
\title{checkLinkPatternsParameters}
\usage{
checkLinkPatternsParameters(ref, alt, context, distance, mutationSymbol,
  reverseComplement, searchPatterns, searchRefHeader, searchAltHeader,
  searchContextHeader, searchIdHeader, searchDistanceHeader,
  searchMutationSymbol, searchReverseComplement, showWarning)
}
\description{
A function to check the parameters of linkPatterns
}
