% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{checkParametersShuffleMutations}
\alias{checkParametersShuffleMutations}
\title{checkParametersShuffleMutations}
\usage{
checkParametersShuffleMutations(dataTable, chromHeader, positionHeader,
  refHeader, altHeader, surroundingHeader, sampleIdHeader, nBootstrap,
  maxDistance, reverseComplement, searchPatterns, searchRefHeader,
  searchAltHeader, searchContextHeader, searchIdHeader,
  searchDistanceHeader, searchReverseComplement, asTibble,
  returnEachBootstrap, searchClusterPatterns, no.cores)
}
\description{
A function to check if the parameters are correct of the
  \code{\link{shuffleMuations}} function
}
