% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(nucleotide, symbols)
}
\arguments{
\item{nucleotide}{A string with a single nucleotide}

\item{symbols}{A data frame with the symbols to match with the nucleotide}
}
\description{
A function to see if the nucleotide is present in the symbol
  data frame
}
