% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{compareContext}
\alias{compareContext}
\title{compareContext}
\usage{
compareContext(searchContext, findContext, mutationSymbol, alphabet,
  searchMutationSymbol)
}
\arguments{
\item{searchContext}{A string with a context (e.g. C.G) This one will act as the
known mutation context for linkPatterns function.}

\item{findContext}{A string with a context (e.g.  C.C)}

\item{mutationSymbol}{A character that stands for the mutation nucleotide
(e.g. "." for C.G)}

\item{alphabet}{A tibble with the nucleotides and their symbols}

\item{searchMutationSymbol}{A string with symbol that stands for the mutated
nucleotide location in the column of the \code{searchContextHeader}. (e.g.
"." in "G.C")}
}
\value{
Boolean if the contexts match or not
}
\description{
Function to compare nucleotides context
}
