% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchClusterPatterns.R
\name{compareNucs}
\alias{compareNucs}
\title{compareNucs}
\usage{
compareNucs(find, search)
}
\arguments{
\item{find}{The nucleotides to be found}

\item{search}{The nucleotides that needs to be matched with}
}
\description{
A function to compare if two sequences are a match.
}
