% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryPatterns.R
\name{convertFactor}
\alias{convertFactor}
\title{convertFactor}
\usage{
convertFactor(x)
}
\arguments{
\item{x}{data frame were the columns with string need to converted to
character}
}
\description{
A function to convert factor columns to character
}
