% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{convertLetter}
\alias{convertLetter}
\title{convertLetter}
\usage{
convertLetter(chars)
}
\arguments{
\item{chars}{List with nucleotides}
}
\description{
Function to get the reverse complement nucleotides
}
