% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyAndAnnotateClusters.R
\name{createIdentTable}
\alias{createIdentTable}
\title{createIdentTable}
\usage{
createIdentTable(dataTable, maxDistance, chromHeader, sampleIdHeader,
  positionHeader, refHeader, altHeader, contextHeader)
}
\description{
Creates the table with extra annotation about the present
  clusters
}
