% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{createPatternFreq}
\alias{createPatternFreq}
\title{createPatternFreq}
\usage{
createPatternFreq(chromHeader, positionHeader, refHeader, altHeader,
  surroundingHeader, sampleIdHeader, maxDistance, shuffleTable,
  linkPatterns, reverseComplement, searchPatterns, searchRefHeader,
  searchAltHeader, searchContextHeader, searchIdHeader,
  searchReverseComplement)
}
\arguments{
\item{chromHeader}{A string; The name of the column with the chromosome nr.}

\item{positionHeader}{A string; The name of the column with the position nr.}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative nucleotide.}

\item{surroundingHeader}{A string with the column header of the surrounding nucleotides.}

\item{sampleIdHeader}{A string; The name of the column with the sample ID.}

\item{maxDistance}{A number; The maximum distance between DNA mutations that
count as clustered.}

\item{linkPatterns}{A boolean to tell if patterns are needed to be found.}

\item{reverseComplement}{A boolean to tell if the ref, alt and context needed
to be searched with the reverse complement.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
data/mutationPatterns.rds is the default search table.}

\item{searchRefHeader}{A string with the column name of the reference
nucleotide of the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the alternative
nucleotide of the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the context
nucleotide of the searchPatterns table.}

\item{searchReverseComplement}{A boolean to also search in the reverse
complement of the searchPatterns tibble. When matched the returned ID will
have [Rev.Com] attached to it.}
}
\description{
A function to create a table with frequencies of patterns.
}
