% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{createRandomMutations}
\alias{createRandomMutations}
\title{createRandomMutations}
\usage{
createRandomMutations(nMut = 500, sampleName = "testSample",
  asTibble = TRUE, sizeSur = 2, refGenomeHg19 = TRUE,
  useChrom = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7",
  "chr8", "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15",
  "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22", "chrX",
  "chrY"))
}
\arguments{
\item{nMut}{The number of mutations that needed to be generated.}

\item{sampleName}{A name for the test sample.}

\item{asTibble}{A Boolean if the returned table needs to be a tibble. Returns
a data.table when FALSE.}

\item{sizeSur}{A number with the ammount of nucleotides left and right of the
mutation. (e.g. sizeSur = 2 --> "CC.GT")}

\item{useChrom}{A vector with the chromosomes that are used. Make sure that
the names are notated as e.g. "chr1".}

\item{refGenome}{A boolean with the prefered reference genome. If TRUE: the
reference will be hg19. if FALSE: HG38.}
}
\value{
A tibble or data.table with random mutation information.
}
\description{
Function to generate random data in a tibble with DNA mutation
  information. For explanation of the columns use cat(comment(<returned
  object>))
}
\examples{
x <- createRandomMutations(nMut = 100,
                           sampleName = "test",
                           sizeSur = 3)

# Proper way of using all the parameters:
createRandomMutations(nMut = 10,
                      sampleName = "name",
                      asTibble = FALSE,
                      sizeSur = 4,
                      refGenomeHg19 = F,
                      useChrom = c("chr1","chr2"))

# See explanation of table columns
cat(comment(x))
}
