% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{createRandomTable}
\alias{createRandomTable}
\title{createRandomTable}
\usage{
createRandomTable(nMut, sampleName, sizeSur, refGenomeHg19, useChrom,
  lenChrom)
}
\arguments{
\item{nMut}{The number of mutations that needed to be generated.}

\item{sampleName}{A name for the test sample.}

\item{sizeSur}{A number with the ammount of nucleotides left and right of the
mutation. (e.g. sizeSur = 2 --> "CC.GT")}

\item{useChrom}{A vector with the chromosomes that are used. Make sure that
the names are notated as e.g. "chr1".}
}
\description{
Function to build the random mutation table and call supporting
  functions.
}
