% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{createShuffleTable}
\alias{createShuffleTable}
\title{createShuffleTable}
\usage{
createShuffleTable(x, chromHeader, positionHeader, refHeader, altHeader,
  surroundingHeader, sampleIdHeader, nBootstrap, showProgress)
}
\arguments{
\item{x}{A table with the reference, alternative and surrounding
nucleotides.}

\item{chromHeader}{A string; The name of the column with the chromosome nr.}

\item{positionHeader}{A string; The name of the column with the position nr.}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative nucleotide.}

\item{surroundingHeader}{A string with the column header of the surrounding nucleotides.}

\item{sampleIdHeader}{A string; The name of the column with the sample ID.}

\item{nBootstrap}{A number with the ammount of bootstraps there need to be excecuted.}

\item{showProgress}{A boolean whether or not to show the progress.}
}
\description{
A function to bootstrap the mutation table
}
