% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{createSummaryPatterns}
\alias{createSummaryPatterns}
\title{createSummaryPatterns}
\usage{
createSummaryPatterns(clusterTable, searchPatterns, searchIdHeader,
  random = FALSE, locationBased = FALSE)
}
\arguments{
\item{clusterTable}{Table with cluster information}

\item{searchPatterns}{A tibble with the known mutation patterns. The
data/mutationPatterns.rds is the default search table.}

\item{grouped}{A Boolean to tell if the data is grouped or not}
}
\description{
A function to create a table with frequencies of patterns.
}
