% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{formatClusterMutations}
\alias{formatClusterMutations}
\title{formatClusterMutations}
\usage{
formatClusterMutations(refs, alts, convert = FALSE)
}
\arguments{
\item{refs}{A list containing the reference nucleotides.}

\item{alts}{A list containing the alternative nucleotides.}

\item{convert}{A boolean that tells if the nucleotides need to be converted
to the reverse complement.}
}
\description{
Function to make a mutation description symbol (e.g. G>C).
}
