% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomDNMs.R
\name{generateRandomDNMs}
\alias{generateRandomDNMs}
\title{GenrateRandomDNMs}
\usage{
generateRandomDNMs(nSamples = 2, nChrom = 2, nDNM = 10)
}
\arguments{
\item{nSamples}{integer; number of samples}

\item{nChrom}{integer; number of chromosomes per sample}

\item{nDNM}{integer; number of mutations per chromosome}
}
\value{
A tibble with DNM information
}
\description{
Function to generate random tibble with DNM information.
For explanation of the columns use cat(comment(<returned object>))
}
