% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{getAlphaMatches}
\alias{getAlphaMatches}
\title{getAlphaMatches}
\usage{
getAlphaMatches(nuc, alphabet)
}
\arguments{
\item{nuc}{string with the nucleotide that is needed to be found}

\item{alphabet}{data frame with nucleotide symbols}
}
\description{
A function to find the symbols that contains the nucleotide
}
