% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterId.R
\name{getClusterId}
\alias{getClusterId}
\title{getClusterId}
\usage{
getClusterId(dataTable, maxDistance, chromHeader = "Chr",
  sampleIdHeader = "sampleID", positionHeader = "Pos")
}
\arguments{
\item{dataTable}{A data.frame or tibble that contains at least chromosome
name, sample ID and position information. The data cannot contain any NA.
For an example use \code{\link{testDataSet}}.}

\item{maxDistance}{A number with the maximum distance between DNA mutations
that are defined as being in a cluster.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. (The data in the column needs to be nummeric.)}
}
\value{
A atomic character vector; contains per index the chromosome nr,
  sample name and any cluster ID's. For example: Chr1 testSample 1
}
\description{
A function that is able to search for clusters in DNA mutation
  table (as explained by argument dataTable) in a single chromosome for a single
  sample.
}
