% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{getNnuc}
\alias{getNnuc}
\title{getNnuc}
\usage{
getNnuc(context1, context2)
}
\arguments{
\item{context1}{surrounding nucleotides of the unknown mutation}

\item{context2}{surrounding nucleotides from the searchPattern table}
}
\value{
Returns a number with amount of nucleotides that has to be compared
}
\description{
A function to determine how many nucleotides has to be compared
}
