% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{getPatternIntersect}
\alias{getPatternIntersect}
\title{getPatternIntersect}
\usage{
getPatternIntersect(clusterList, patternHeader)
}
\arguments{
\item{clusterList}{A tibble with the mutations information.}

\item{patternHeader}{A string with the column header of the patterns.}
}
\description{
A function to find the intersect of patterns between mutations.
}
