% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{getReverseComplement}
\alias{getReverseComplement}
\title{getReverseComplement}
\usage{
getReverseComplement(x)
}
\description{
A function to get the reverse complement of the nucleotides in param x
}
