% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearchPatterns.R
\name{getSearchPatterns}
\alias{getSearchPatterns}
\title{getSearchPatterns}
\usage{
getSearchPatterns(reverse = TRUE, renameReverse = FALSE,
  asTibble = TRUE)
}
\arguments{
\item{reverse}{A Boolean if the reverse complementend also needed to be
added.}

\item{renameReverse}{A Boolean if the id of the process needs to be renamed.
This has the effect on the cMut functions that it will no longer treat the
reverse complement and non reverse complement as the same.}

\item{asTibble}{A Boolean if the returned table needs to be a tibble. Else it
will sent a data.frame.}
}
\description{
A function to get the default search pattern table
}
\note{
Please note that if there are patterns where the reverse complement is
  the same as the original sequence then the reverse complement won't be
  made. This is done to avoid double counting when \code{renameReverse =
  TRUE} is used.
}
\seealso{
See \code{\link{mutationPatterns}} help page for full description
  about the columns and values.
}
