% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{getSurrounding}
\alias{getSurrounding}
\title{getSurrounding}
\usage{
getSurrounding(chr, loc, table, lenChrom)
}
\arguments{
\item{chr}{A string with the chromosome name}

\item{loc}{A number with the mutation location}

\item{table}{A table containing the mutation information}

\item{lenChrom}{A vector with chromosomes length}
}
\description{
Function to get the surrounding of a mutation
}
