% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{identifyClusters}
\alias{identifyClusters}
\title{identifyClusters}
\usage{
identifyClusters(x, maximumDistance, chromHeader = "Chr",
  sampleIdHeader = "sampleID", positionHeader = "Pos")
}
\arguments{
\item{x}{A tibble that contains at least chromosome nr., sampleID and position information.
The data cannot contain any NA.}

\item{chromHeader}{A string; The name of the column with the chromosome nr.}

\item{positionHeader}{A string; The name of the column with the position nr.}
}
\value{
A atomic character vector;
contains per index the chromosoom nr, sample naam
and any cluster ID's. For example: Chr1 15-15-15 c(10,50,100)
}
\description{
A function that is able to search for clusters in DNA mutation table (as explained by argument x)
in a single chromosoom for a single sample.
}
