% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{linkPatterns}
\alias{linkPatterns}
\title{linkPatterns}
\usage{
linkPatterns(ref, alt, context, distance = NULL, mutationSymbol = ".",
  reverseComplement = FALSE, searchPatterns = mutationPatterns,
  searchRefHeader = "ref", searchAltHeader = "alt",
  searchContextHeader = "surrounding", searchIdHeader = "process",
  searchDistanceHeader = "maxDistance", searchMutationSymbol = ".",
  searchReverseComplement = TRUE, showWarning = TRUE)
}
\arguments{
\item{ref}{A string containing the reference nucleotide. (e.g. "C")}

\item{alt}{A string containing the alternative nucleotide. (e.g. "T")}

\item{context}{A string containing the surrounding nucleotides. (e.g. "G.C")}

\item{distance}{A number that tells the distance to the nearest mutation in a
cluster. This parameter should stay NULL unless the \code{searchPatterns}
table has a distance column and it's needed to be take into acount while
comparing.}

\item{mutationSymbol}{A string with the symbol that stands for the mutated
nucleotide location in the \code{context}. (e.g. "." in "G.C")}

\item{reverseComplement}{A Boolean to tell if the \code{ref}, \code{alt} and
\code{context} needed to be searched with the reverse complement.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchRefHeader}{A string with the column name of the one with the
reference nucleotide in the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the one with the
alternative nucleotide in the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the one with the
context nucleotide in the searchPatterns table.}

\item{searchDistanceHeader}{A string with the column name of the one with the
maximum distance between clustered mutations. Not needed if the distance
parameter is NULL. NA's within this column are allowed.}

\item{searchMutationSymbol}{A string with symbol that stands for the mutated
nucleotide location in the column of the \code{searchContextHeader}. (e.g.
"." in "G.C")}

\item{searchReverseComplement}{A boolean to also search the patterns in the
reverse complement of the searchPatterns tibble.}

\item{showWarning}{A Boolean if warnings about the parameters are alowed.}

\item{searchSource}{A string with the column name of the ID of the known
pattern.}
}
\value{
list with the matched patterns. If nothing's found, return
  an empty list.
}
\description{
A function that is able to find matches between a submitted
  unkown mutation and a known mutation pattern table.
}
\examples{
results <- linkPatterns("C","T","TA.T")

# To see the default searchPattern table and it's information:
?mutationPatterns
mutationPatterns
}
