% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationPatterns.R
\docType{data}
\name{mutationPatterns}
\alias{mutationPatterns}
\title{Mutation patterns of mutators}
\format{A data frame with 10 rows and 4 columns
\describe{
  \item{process}{process that is connected with the mutation pattern}
  \item{ref}{nucleotide that is found in the reference genome}
  \item{alt}{nucleotide that the reference is transformed in}
  \item{surrounding}{the surrounding nucleotides where the mutation takes place}
}}
\usage{
mutationPatterns
}
\description{
A small data set of mutationpatterns of known mutators in human DNA.
}
\keyword{datasets}
