% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{nucToSymbol}
\alias{nucToSymbol}
\title{nucToSymbol}
\usage{
nucToSymbol(nuc, mutationSymbol)
}
\arguments{
\item{nuc}{vector with the nucleotides}

\item{mutationSymbol}{symbol that represent the mutation site}
}
\description{
Function to convert a vector nucleotides to a single symbol
}
