% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchClusterPatterns.R
\name{searchClusterPatterns}
\alias{searchClusterPatterns}
\title{searchClusterPatterns}
\usage{
searchClusterPatterns(groupedClusters, searchPatterns = mutationPatterns,
  searchRefHeader = "ref", searchAltHeader = "alt",
  searchDistanceHeader = "maxDistance", searchIdHeader = "process")
}
\arguments{
\item{groupedClusters}{Table gained by using the groupClusters function.}
}
\description{
searchClusterPatterns
}
