% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearchPatterns.R
\name{setdiff.data.frame}
\alias{setdiff.data.frame}
\title{setdiff.data.frame}
\usage{
setdiff.data.frame(x, y)
}
\arguments{
\item{x}{data frame that is used to return the differences from y}

\item{y}{data frame that is used to compare with x}
}
\description{
Quick function to apply \code{\link{setdiff}} on dataframes
}
