% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{shuffleMutations}
\alias{shuffleMutations}
\title{shuffleMutations}
\usage{
shuffleMutations(x, chromHeader = "chrom", positionHeader = "start",
  refHeader = "ref", altHeader = "alt",
  surroundingHeader = "surrounding", sampleIdHeader = "sampleIDs",
  nBootstrap = 1000, maxDistance = 20000, reverseComplement = FALSE,
  searchPatterns = NULL, searchRefHeader = "ref",
  searchAltHeader = "alt", searchContextHeader = "surrounding",
  searchIdHeader = "proces", searchReverseComplement = TRUE,
  tibble = TRUE)
}
\arguments{
\item{x}{A table with the reference, alternative and surrounding
nucleotides.}

\item{chromHeader}{A string; The name of the column with the chromosome nr.}

\item{positionHeader}{A string; The name of the column with the position nr.}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative nucleotide.}

\item{surroundingHeader}{A string with the column header of the surrounding nucleotides.}

\item{sampleIdHeader}{A string; The name of the column with the sample ID.}

\item{nBootstrap}{A number with the ammount of bootstraps there need to be excecuted.}

\item{maxDistance}{A number; The maximum distance between DNA mutations that
count as clustered.}

\item{reverseComplement}{A boolean to tell if the ref, alt and context needed
to be searched with the reverse complement.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
data/mutationPatterns.rds is the default search table.}

\item{searchRefHeader}{A string with the column name of the reference
nucleotide of the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the alternative
nucleotide of the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the context
nucleotide of the searchPatterns table.}

\item{searchReverseComplement}{A boolean to also search in the reverse
complement of the searchPatterns tibble. When matched the returned ID will
have [Rev.Com] attached to it.}

\item{tibble}{A boolean if the table has to be a tibble}
}
\description{
A function to shuffle the reference, alternative and surrounding
  nucleotides. Usefull to determine the chance that the results of your data
  might be due randomness.
}
