% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{shuffleParallel}
\alias{shuffleParallel}
\title{shuffleParallel}
\usage{
shuffleParallel(dataTable, chromHeader, positionHeader, refHeader,
  altHeader, surroundingHeader, sampleIdHeader, nBootstrap, maxDistance,
  reverseComplement, searchPatterns, searchRefHeader, searchAltHeader,
  searchContextHeader, searchIdHeader, searchDistanceHeader,
  searchReverseComplement, searchClusterPatterns, no.cores, resultTable)
}
\description{
Function to preform the shuffling and annotating in parallel
}
