% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{summarizeBootstrap}
\alias{summarizeBootstrap}
\title{summarizeBootstrap}
\usage{
summarizeBootstrap(resultTable, resultTables, searchIdHeader)
}
\arguments{
\item{resultTable}{A dataframe were the results of each bootstrap is put into}

\item{resultTables}{List with all the results of the bootstrap}

\item{searchIdHeader}{A string with the column name of the one with the
pattern IDs.}
}
\description{
A function to summarize the bootstrap results
}
