% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{addClusterTableComment}
\alias{addClusterTableComment}
\title{addClusterTableComment}
\usage{
addClusterTableComment(table, patternIntersect, clusterIdHeader,
  searchClusterPatterns)
}
\description{
Adds explanation to the table about the columns
}
