% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyAndAnnotateClusters.R
\name{addDistance}
\alias{addDistance}
\title{addDistance}
\usage{
addDistance(ranges, maxDistance)
}
\arguments{
\item{ranges}{A GRange object which were created during identifyAndAnnotateClusters function}

\item{maxDistance}{A integer; The maximum distance between DNMs that count as clustered.}
}
\value{
A GRange with added distance and a logical column as metadata
}
\description{
A function that adds distances to nearest DNM to the tibble of the identifyAndAnnotateClusters function
}
