% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{addIdentTableComment}
\alias{addIdentTableComment}
\title{addIdentTableComment}
\usage{
addIdentTableComment(dataTable, linkPatterns)
}
\arguments{
\item{dataTable}{A data.frame or tibble that contains at least chromosome
name, sample ID and position information. The data cannot contain any NA.
For an example use \code{\link{testDataSet}}.}

\item{linkPatterns}{Boolean if pattern columns are added}
}
\description{
Adds commentary to the results of the
  \code{\link{identifyClusters}} function.
}
