% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyAndAnnotateClusters.R
\name{addLinkPatterns}
\alias{addLinkPatterns}
\title{addLinkPatterns}
\usage{
addLinkPatterns(x, refHeader, altHeader, contextHeader, mutationSymbol,
  reverseComplement, searchPatterns, searchRefHeader, searchAltHeader,
  searchContextHeader, searchIdHeader, searchReverseComplement,
  checkHeader)
}
\arguments{
\item{x}{A tibble that contains at least chromosome nr., sampleID and
position information. The data cannot contain any NA. For an example use
createRandomMutations function.}

\item{refHeader}{Contains the name of the column with the reference
nucleotides..}

\item{altHeader}{Contains the name of the column with the alternative
nucleotides.}

\item{contextHeader}{A String; The name of the column with the context.}

\item{mutationSymbol}{A string with the symbol that stands for the mutated
nucleotide location (e.g. "." in G.C).}

\item{reverseComplement}{A boolean to tell if the ref, alt and context needed
to be searched with the reverse complement.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
data/mutationPatterns.rds is the default search table.}

\item{searchRefHeader}{A string with the column name of the reference
nucleotide of the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the alternative
nucleotide of the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the context
nucleotide of the searchPatterns table.}

\item{searchReverseComplement}{A boolean to also search in the reverse
complement of the searchPatterns tibble. When matched the returned ID will
have [Rev.Com] attached to it.}
}
\description{
addLinkPatterns
}
