% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{addLinkPatterns}
\alias{addLinkPatterns}
\title{addLinkPatterns}
\usage{
addLinkPatterns(table, refHeader, altHeader, contextHeader, mutationSymbol,
  reverseComplement, searchPatterns, searchRefHeader, searchAltHeader,
  searchContextHeader, searchIdHeader, searchReverseComplement,
  searchMutationSymbol, checkHeader, searchDistanceHeader,
  linkClustersOnly, renameReverse)
}
\arguments{
\item{table}{table with the clusters information}

\item{refHeader}{Contains the name of the column with the reference
nucleotides.}

\item{altHeader}{Contains the name of the column with the alternative
nucleotides.}

\item{contextHeader}{A string with the name of the column with the context.
The data inside this column is e.g. "C.G" hereby stands the "." for the
location of the mutation. What symbol is used to describe this location is
arbitrary but be sure to adjust the \code{mutationSymbol} accordingly when
searching for patterns. The \code{contextHeader} is irrelevant if
\code{linkPatterns} is FALSE.}

\item{mutationSymbol}{A string with the symbol that stands for the mutated
nucleotide location in the \code{context}. (e.g. "." in "G.C")}

\item{reverseComplement}{A Boolean to tell if the \code{ref}, \code{alt} and
\code{context} needed to be searched with the reverse complement. Irrelevant
if \code{searchReverseComplement = TRUE}.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchRefHeader}{A string with the column name of the one with the
reference nucleotide in the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the one with the
alternative nucleotide in the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the one with the
context nucleotide in the searchPatterns table.}

\item{searchIdHeader}{A string with the column name of the one with the
pattern IDs.}

\item{searchReverseComplement}{A boolean to also search the patterns in the
reverse complement of the searchPatterns tibble.}

\item{searchMutationSymbol}{A string with symbol that stands for the mutated
nucleotide location in the column of the \code{searchContextHeader}. (e.g.
"." in "G.C")}

\item{checkHeader}{name of the Boolean check column}

\item{searchDistanceHeader}{A string with the column name of the one with the
maximum distance between clustered mutations. Not needed if the distance
parameter is NULL. NA's within this column are allowed.}

\item{linkClustersOnly}{A boolean to tell if only the clustered mutations are
needed to be linked with the patterns in the \code{searchPatterns} table.
When it is FALSE all the mutations will be used.}

\item{renameReverse}{A Boolean if the id of the process needs to be renamed.
This has the effect on the cMut functions that it will no longer treat the
reverse complement and non reverse complement as the same. This parameter
will irrelevant if \code{searchReverseComplement} is FALSE.}
}
\description{
A function to add columns to the sent table with information
  gathered from the \code{\link{linkPatterns}} function.
}
