% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{addShuffleMutationComment}
\alias{addShuffleMutationComment}
\title{addShuffleMutationComment}
\usage{
addShuffleMutationComment(results, searchIdHeader)
}
\description{
A function to add comment about the result table.
}
