% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{callSearchclusterPatterns}
\alias{callSearchclusterPatterns}
\title{callSearchclusterPatterns}
\usage{
callSearchclusterPatterns(table, patternIntersect, searchPatterns,
  searchRefHeader, searchAltHeader, searchIdHeader, searchDistanceHeader,
  searchReverseComplement, renameReverse)
}
\description{
A function to add and change columns of the groupClusters
  function with the results of the \code{\link{searchClusterPatterns}}
  function.
}
