% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{compareContext}
\alias{compareContext}
\title{compareContext}
\usage{
compareContext(context2, context1, mutationSymbol, alphabet,
  searchMutationSymbol)
}
\arguments{
\item{context2}{A string with a context (e.g. C.G) This one will act as the
known mutation context for linkPatterns function.}

\item{context1}{A string with a context (e.g.  C.C)}

\item{mutationSymbol}{A character that stands for the mutation nucleotide
(e.g. "." for C.G)}

\item{alphabet}{A tibble with the nucleotides and their symbols}
}
\value{
Boolean if the contexts match or not
}
\description{
Function to compare nucleotides context
}
