% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{createGroupTable}
\alias{createGroupTable}
\title{createGroupTable}
\usage{
createGroupTable(table, clusterIdHeader, refHeader, altHeader,
  contextHeader, mutationSymbol, showWarning)
}
\description{
A function to group the table by the cluster IDs and add
  annotation to it
}
