% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{createIdentTable}
\alias{createIdentTable}
\title{createIdentTable}
\usage{
createIdentTable(dataTable, maxDistance, chromHeader, sampleIdHeader,
  positionHeader, refHeader, altHeader, contextHeader)
}
\arguments{
\item{dataTable}{A data.frame or tibble that contains at least chromosome
name, sample ID and position information. The data cannot contain any NA.
For an example use \code{\link{testDataSet}}.}

\item{maxDistance}{A number with the maximum distance between DNA mutations
that are defined as being in a cluster.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. (The data in the column needs to be nummeric.)}

\item{refHeader}{Contains the name of the column with the reference
nucleotides.}

\item{altHeader}{Contains the name of the column with the alternative
nucleotides.}

\item{contextHeader}{A string with the name of the column with the context.
The data inside this column is e.g. "C.G" hereby stands the "." for the
location of the mutation. What symbol is used to describe this location is
arbitrary but be sure to adjust the \code{mutationSymbol} accordingly when
searching for patterns. The \code{contextHeader} is irrelevant if
\code{linkPatterns} is FALSE.}
}
\description{
Creates the table with extra annotation about the present
  clusters
}
