% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{createRandomMutations}
\alias{createRandomMutations}
\title{createRandomMutations}
\usage{
createRandomMutations(nMut = 500, sampleName = "testSample")
}
\arguments{
\item{nMut}{number of mutations that needed to be generated}

\item{sampleName}{A name for the test sample}
}
\value{
A tibble with mutation information
}
\description{
Function to generate random data in a tibble with DNA mutation information.
For explanation of the columns use cat(comment(<returned object>))
}
