% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{createRandomMutations}
\alias{createRandomMutations}
\title{createRandomMutations}
\usage{
createRandomMutations(nMut = 500, sampleName = "testSample",
  tibble = TRUE, sizeSur = 2, refGenomeHg19 = TRUE)
}
\arguments{
\item{nMut}{The number of mutations that needed to be generated.}

\item{sampleName}{A name for the test sample.}

\item{tibble}{A Boolean if the returned table needs to be a tibble.}

\item{sizeSur}{A number with the ammount of nucleotides left and right of the
mutation. (e.g. sizeSur = 2 --> "CC.GT")}

\item{refGenome}{A boolean with the prefered reference genome. If TRUE: the reference will be hg19. if FALSE: HG38.}
}
\value{
A tibble with mutation information
}
\description{
Function to generate random data in a tibble with DNA mutation
  information. For explanation of the columns use cat(comment(<returned
  object>))
}
\examples{
x <- createRandomMutations(nMut = 100,
                           sampleName = "test",
                           sizeSur = 3)

# See explanation of table columns
cat(comment(x))
}
