% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{createShuffleTable}
\alias{createShuffleTable}
\title{createShuffleTable}
\usage{
createShuffleTable(x, chromHeader, positionHeader, refHeader, altHeader,
  surroundingHeader, sampleIdHeader)
}
\arguments{
\item{x}{A table with the reference, alternative and surrounding nucleotides.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. The data in the column needs to be numberic.}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative
nucleotide.}

\item{surroundingHeader}{A string with the column header of the surrounding
nucleotides.}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}
}
\description{
A function to bootstrap the mutation table
}
