% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryPatterns.R
\name{createSummaryPatterns}
\alias{createSummaryPatterns}
\title{createSummaryPatterns}
\usage{
createSummaryPatterns(clusterTable, searchPatterns, searchIdHeader)
}
\arguments{
\item{clusterTable}{Table with cluster information}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchIdHeader}{A string with the column name of the one with the
pattern IDs.}
}
\description{
A function to create a table with frequencies of patterns.
}
