% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaAlphabet.R
\docType{data}
\name{dnaAlphabet}
\alias{dnaAlphabet}
\title{DNA Alphabet}
\format{A data frame with 16 rows and 2 columns:
\describe{
  \item{symbol}{the symbol that is used to describe certain nucleotides}
  \item{represent}{the represented nucleotides}
}}
\usage{
dnaAlphabet
}
\description{
A dataset with the DNA Alphabet
}
\keyword{datasets}
