% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{fuseSurroundings}
\alias{fuseSurroundings}
\title{fuseSurroundings}
\usage{
fuseSurroundings(surroundings, mutationSymbol, showWarning)
}
\description{
A function to combine surroundings by using nucleotide symbols.
  See \code{\link{dnaAlphabet}} for an explanation what these symbols are.
}
