% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomDNMs.R
\name{generateRandomDNMs}
\alias{generateRandomDNMs}
\title{GenrateRandomDNMs}
\usage{
generateRandomDNMs(nDNM = 100, sampleName = "testSample")
}
\arguments{
\item{nDNM}{number of mutations that needed to be generated}

\item{sampleName}{A name for the test sample}
}
\value{
A tibble with DNM information
}
\description{
Function to generate random data in a tibble with DNM information.
For explanation of the columns use cat(comment(<returned object>))
}
