% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{getClusterType}
\alias{getClusterType}
\title{getClusterType}
\usage{
getClusterType(plusStrand, minusStrand)
}
\arguments{
\item{plusStrand}{A string with the plusStrand mutation description.}

\item{minusStrand}{A string with the minusStrand mutation description.}
}
\description{
Function to sort and create a cluster type (e.g. C>G C>G / G>C
  G>C).
}
