% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{getRef}
\alias{getRef}
\title{getRef}
\usage{
getRef(posTable, refGenomeHg19)
}
\arguments{
\item{posTable}{a table with chromosome, start and stop position information}

\item{refGenomeHg19}{A Boolean to tell if the preferred reference genome has
to be Hg19. Hg38 will be used when \code{refGenome = FALSE}.}
}
\value{
A string with the reference nucleotides.
}
\description{
A function to get the reference nucleotide.
}
