% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomMutations.R
\name{getRef}
\alias{getRef}
\title{getRef}
\usage{
getRef(x)
}
\arguments{
\item{x}{string that contains the chromosome name, start and stop location and is seperated by "-"}
}
\description{
A function to get the reference nucleotide
}
