% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{getRef}
\alias{getRef}
\title{getRef}
\usage{
getRef(x, lenChrom, refGenomeHg19)
}
\arguments{
\item{x}{A string that contains the chromosome name, start and stop location
and is separated by "-".}

\item{lenChrom}{A vector with chromosomes length}
}
\value{
A string with the reference nucleotides.
}
\description{
A function to get the reference nucleotide.
}
