% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{getRefData}
\alias{getRefData}
\title{getRefData}
\usage{
getRefData(chr, pos, sizeSur, lenChrom)
}
\arguments{
\item{chr}{A string with the chromosome name}

\item{pos}{A number with the mutation location}

\item{sizeSur}{A number with the amount of nucleotides there has to be
arround the mutation.}

\item{lenChrom}{A vector with chromosomes length}

\item{table}{A table containing the mutation information}
}
\description{
Function to get the surrounding of a mutation
}
