% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomMutations.R
\name{getRefData}
\alias{getRefData}
\title{getRefData}
\usage{
getRefData(data, sizeSur, lenChrom, refGenomeHg19)
}
\arguments{
\item{data}{Table with the chromosome, start information}

\item{sizeSur}{A number with the amount of nucleotides there has to be
around the mutation.}

\item{lenChrom}{vector with the different size of the chromosomes}

\item{refGenomeHg19}{A Boolean to tell if the preferred reference genome has
to be Hg19. Hg38 will be used when \code{refGenome = FALSE}.}
}
\value{
A string with the surrounding and mutated position reference
  nucleotides of a mutation. Each part separated with "-"
}
\description{
Function to get the reference nucleotides of a mutation.
}
