% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearchPatterns.R
\name{getRevComTable}
\alias{getRevComTable}
\title{getRevComTable}
\usage{
getRevComTable(table, refHeader, altHeader, contextHeader = NULL,
  idHeader)
}
\description{
A function to get the reverse complement of the known mutations
}
