% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPatterns.R
\name{getReverseComplement}
\alias{getReverseComplement}
\title{getReverseComplement}
\usage{
getReverseComplement(x)
}
\arguments{
\item{x}{nucleotides that need to be converted to reverse complement}
}
\description{
A function to get the reverse complement of a sequence
}
