% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearchPatterns.R
\name{getSearchPatterns}
\alias{getSearchPatterns}
\title{getSearchPatterns}
\usage{
getSearchPatterns(reverse = TRUE, asTibble = T)
}
\arguments{
\item{reverse}{A Boolean if the reverse complementend also needed to be added}
}
\description{
A function to get the default search pattern table
}
