% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryPatterns.R
\name{getSummaryPatterns}
\alias{getSummaryPatterns}
\title{getSummaryPatterns}
\usage{
getSummaryPatterns(groupedClusters, searchPatterns = NULL,
  searchIdHeader = "process", renameReverse = FALSE, asTibble = TRUE)
}
\arguments{
\item{groupedClusters}{A table generated from the
\code{\link{groupClusters}}.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchIdHeader}{A string with the column name of the one with the
pattern IDs.}

\item{renameReverse}{A Boolean to tell if the pattern IDs from the
searchPatterns also needs to be used to find the patterns with "
[Rev.Com.]". If FALSE and those patterns are available, then they will be
counted in the Unidentified row.}

\item{asTibble}{A Boolean to tell if the returned table needs to be a tibble
or a data.frame.}
}
\description{
A optional follow-up function to summarize the results of the
  \code{foundPatterns} column from \code{\link{groupClusters}}.
}
\note{
If the \code{groupedClusters} table contains patterns that are not
  present in the \code{searchPattern} table, then they will be marked as
  \code{Unidentified} together with clusters without patterns.
}
\examples{
# Example dataset
data <- testDataSet

# Use the following functions to get the necessary table
results <- identifyAndAnnotateClusters(dataTable    = data,
                                       maxDistance  = 20000,
                                       linkPatterns = TRUE)
groupResults <- groupClusters(dataTable             = results,
                              searchClusterPatterns = TRUE,
                              patternIntersect      = TRUE)

# Use the getSummaryPatterns function to see the summary of found patterns
summary <- getSummaryPatterns(groupResults)
summary

# For more information about the columns use:
cat(comment(summary))
}
