% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryPatterns.R
\name{getSummaryPatterns}
\alias{getSummaryPatterns}
\title{getSummaryPatterns}
\usage{
getSummaryPatterns(groupedClusters, searchPatterns = NULL,
  searchIdHeader = "process", searchReverseComplement = TRUE)
}
\arguments{
\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}
}
\description{
A function to summarise the found patterns
}
\examples{
data <- testDataSet
results <- identifyAndAnnotateclusters(data,20000,linkPatterns == T)
}
