% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClusters.R
\name{groupClusters}
\alias{groupClusters}
\title{groupClusters}
\usage{
groupClusters(table, clustIdHeader = "clusterId", refHeader = "ref",
  altHeader = "alt")
}
\arguments{
\item{table}{A table with columns containing cluster ID's, reference and alternative nucleotide}

\item{refHeader}{Contains the name of the column with the reference nucleotides}

\item{altHeader}{Contains the name of the column with the alternative nucleotides}

\item{clusterIdHeader}{Contains the name of the column with the cluster IDs}
}
\description{
A function that will group the clusters and give the mutation consensus back
}
\examples{
# Example of a table containing the right columns and data
test <- createRandomMutations()
test <- identifyAndAnnotateClusters(test, 20000, chromHeader = "chrom",
                                    sampleIdHeader = "sampleIDs",positionHeader = "start")
}
