% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyAndAnnotateClusters.R
\name{identifyAndAnnotateClusters}
\alias{identifyAndAnnotateClusters}
\title{identifyAndAnnotateClusters}
\usage{
identifyAndAnnotateClusters(x, maxDistance, chromHeader = "Chr",
  sampleIdHeader = "sampleID", positionHeader = "Pos")
}
\arguments{
\item{x}{A tibble that contains at least chromosome nr., sampleID and position information.
The data cannot contain any NA.}

\item{maxDistance}{A number; The maximum distance between DNA mutations that count as clustered.}

\item{chromHeader}{A string; The name of the column with the chromosome nr.}

\item{positionHeader}{A string; The name of the column with the position nr.}

\item{sampleIDHeader}{A string; The name of the column with the sample ID.}
}
\value{
The tibble that was sent as an argument for this fuction with extra columns:
clusterId, is.clustered and distance till nearest mutation below the maximum distance.
}
\description{
A function that finds and annotate clusters in a genomic data tibble.
}
