% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyClusters.R
\name{identifyClusters}
\alias{identifyClusters}
\title{identifyClusters}
\usage{
identifyClusters(x, maxDistance, chromHeader = "Chr",
  sampleIdHeader = "sampleID", positionHeader = "Pos")
}
\arguments{
\item{maxDistance}{A number; The maximum distance between DNA mutations that
count as clustered.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. The data in the column needs to be numberic.}
}
\value{
A atomic character vector; contains per index the chromosoom nr,
  sample naam and any cluster ID's. For example: Chr1 testSample 1
}
\description{
A function that is able to search for clusters in DNA mutation
  table (as explained by argument x) in a single chromosoom for a single
  sample.
}
