% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchClusterPatterns.R
\name{searchClusterPatterns}
\alias{searchClusterPatterns}
\title{searchClusterPatterns}
\usage{
searchClusterPatterns(groupedClusters, searchPatterns = mutationPatterns,
  searchRefHeader = "ref", searchAltHeader = "alt",
  searchDistanceHeader = "maxDistance", searchIdHeader = "process")
}
\arguments{
\item{groupedClusters}{Table gained by using the \code{\link{groupClusters}}
function.}
}
\description{
A function to search cluster patterns based on reference,
  variant and distance of the nucleotides within the cluster.
}
