% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{shuffleMutations}
\alias{shuffleMutations}
\title{shuffleMutations}
\usage{
shuffleMutations(x, chromHeader = "chrom", positionHeader = "start",
  refHeader = "ref", altHeader = "alt",
  surroundingHeader = "surrounding", sampleIdHeader = "sampleIDs",
  nBootstrap = 1000, maxDistance = 20000, reverseComplement = FALSE,
  searchPatterns = NULL, searchRefHeader = "ref",
  searchAltHeader = "alt", searchContextHeader = "surrounding",
  searchIdHeader = "process", searchReverseComplement = TRUE,
  tibble = TRUE, saveEachBootstrap = FALSE, saveFileName = NULL,
  searchClusterPatterns = FALSE)
}
\arguments{
\item{x}{A table with the reference, alternative and surrounding nucleotides.
The best data to use is the output of the
\code{\link{identifyAndAnnotateClusters}} where \code{is.clustered} is
TRUE.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. The data in the column needs to be numberic.}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative
nucleotide.}

\item{surroundingHeader}{A string with the column header of the surrounding
nucleotides.}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}

\item{nBootstrap}{A number with the ammount of bootstraps there need to be
excecuted.}

\item{maxDistance}{A number; The maximum distance between DNA mutations that
count as clustered.}

\item{reverseComplement}{A Boolean to tell if the \code{ref}, \code{alt} and
\code{context} needed to be searched with the reverse complement.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchRefHeader}{A string with the column name of the one with the
reference nucleotide in the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the one with the
alternative nucleotide in the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the one with the
context nucleotide in the searchPatterns table.}

\item{searchReverseComplement}{A boolean to also search the patterns in the
reverse complement of the searchPatterns tibble.}

\item{saveEachBootstrap}{A Boolean if the summaries per bootstrap are needed
to be saved}

\item{saveFileName}{A string with the name and location of the file that
contains the idivudual summaries. Only needed if saveEachBootstrap is TRUE.}
}
\description{
A function to shuffle the reference, alternative and surrounding
  nucleotides. Then it will use the \code{\link{identifyAndAnnotateClusters}}
  and \code{\link{groupClusters}} functions and returns a summary of the
  frequency of found patterns.
}
\examples{
identResults <- identifyAndAnnotateClusters(testDataSet,20000, linkPatterns = T)

# If only the mutation patterns are needed searched:
shuffleResults <- shuffleMutations(identResults,nBootstrap = 5)

# If also the cluster patterns are needed to be added:
shuffleResults <- shuffleMutations(identResults,
                                   nBootstrap = 5,
                                   searchClusterPatterns = T)

}
