% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{shuffleMutations}
\alias{shuffleMutations}
\title{shuffleMutations}
\usage{
shuffleMutations(dataTable, chromHeader = "chrom",
  positionHeader = "start", refHeader = "ref", altHeader = "alt",
  contextHeader = "surrounding", sampleIdHeader = "sampleIDs",
  nBootstrap = 1000, maxDistance = 20000, linkPatterns = TRUE,
  reverseComplement = FALSE, searchPatterns = NULL,
  searchRefHeader = "ref", searchAltHeader = "alt",
  searchContextHeader = "surrounding", searchIdHeader = "process",
  searchDistanceHeader = "maxDistance", searchReverseComplement = TRUE,
  asTibble = TRUE, returnEachBootstrap = FALSE,
  searchClusterPatterns = TRUE, renameReverse = FALSE,
  no.cores = parallel::detectCores())
}
\arguments{
\item{dataTable}{A table with the reference, alternative and surrounding
nucleotides. The best data to use is the output of the
\code{\link{identifyClusters}} where \code{is.clustered} is
TRUE.}

\item{chromHeader}{A string with the name of the column with the chromosome
name. (So the data in the column needs to be notated as e.g. "chr2")}

\item{positionHeader}{A string with the name of the column with the position
of the mutation. (The data in the column needs to be nummeric.)}

\item{refHeader}{A string with the column header of the reference nucleotide.}

\item{altHeader}{A string with the column header of the alternative
nucleotide.}

\item{contextHeader}{A string with the column header of the context
nucleotides.}

\item{sampleIdHeader}{A string with the name of the column with the sample
ID.}

\item{nBootstrap}{A number with the amount of bootstraps there need to be
executed.}

\item{maxDistance}{A number with the maximum distance between DNA mutations
that are defined as being in a cluster.}

\item{linkPatterns}{A Boolean to tell if it's necessary to try and link the
mutations to patterns. If FALSE then the \code{search...} parameters and
the \code{linkClustersOnly} parameter are irrelevant. For more information
see \code{\link{linkPatterns}}.}

\item{reverseComplement}{A Boolean to tell if the \code{ref}, \code{alt} and
\code{context} needed to be searched with the reverse complement. Irrelevant
if \code{searchReverseComplement = TRUE}.}

\item{searchPatterns}{A tibble with the known mutation patterns. The
\code{\link{mutationPatterns}} is the default search table.}

\item{searchRefHeader}{A string with the column name of the one with the
reference nucleotide in the searchPatterns table.}

\item{searchAltHeader}{A string with the column name of the one with the
alternative nucleotide in the searchPatterns table.}

\item{searchContextHeader}{A string with the column name of the one with the
context nucleotide in the searchPatterns table.}

\item{searchIdHeader}{A string with the column name of the one with the
pattern IDs.}

\item{searchDistanceHeader}{A string with the column name of the one with the
maximum distance between clustered mutations. Not needed if the distance
parameter is NULL. NA's within this column are allowed.}

\item{searchReverseComplement}{A boolean to also search the patterns in the
reverse complement of the searchPatterns tibble.}

\item{asTibble}{A Boolean if the returned results needs to be a tibble. It
will return a data.frame otherwise. Irrelevant if \code{returnEachBootstrap
== TRUE}}

\item{returnEachBootstrap}{A Boolean if the summaries per bootstrap are
needed to be returned. If TRUE then the return value will be a list with a
summary of each bootstrap.}

\item{searchClusterPatterns}{A Boolean if it's needed to search to cluster
patterns (e.g. GA > TT).}

\item{renameReverse}{A Boolean if the id of the process needs to be renamed.
This has the effect on the cMut functions that it will no longer treat the
reverse complement and non reverse complement as the same. This parameter
will irrelevant if \code{searchReverseComplement} is FALSE.}

\item{no.cores}{A number with the amount of clusters that is allowed to use
during shuffle. Default is maximum amount of cores present on the system.}
}
\description{
A function to shuffle the reference, alternative and surrounding
  nucleotides. Then it will use the \code{\link{identifyClusters}}
  and \code{\link{groupClusters}} functions and returns a summary of the
  frequency of found patterns.
}
\examples{
identResults <- identifyClusters(dataTable    = cMut::testDataSet,
                                 maxDistance  = 20000,
                                 linkPatterns = TRUE)
clusteredMutations <- identResults[identResults$is.clustered, ]

# If only the mutation patterns are needed searched:
\donttest{shuffleResults <- shuffleMutations(dataTable             = clusteredMutations,
                                   searchClusterPatterns = FALSE,
                                   no.cores              = 2)}

# If also the cluster patterns are needed to be added:
\donttest{shuffleResults <- shuffleMutations(dataTable  = identResults[identResults$is.clustered,],
                                    no.cores   = 2)}
}
