% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleMutations.R
\name{summarizeBootstrap}
\alias{summarizeBootstrap}
\title{summarizeBootstrap}
\usage{
summarizeBootstrap(resultTable, resultTables, searchIdHeader)
}
\description{
A function to summarize the bootstrap results
}
