% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDataSet.R
\docType{data}
\name{testDataSet}
\alias{testDataSet}
\title{A test dataset}
\format{A data frame with 1867 rows and 7 columns:
\describe{
  \item{chrom}{Chromosome where the mutation is localized}
  \item{start}{Start location of the mutation}
  \item{stop}{Stop location of the mutation}
  \item{ref}{Reference nucleotide}
  \item{alt}{The outcome nucleotide of the mutation}
  \item{sampleIDs}{sample name}
  \item{surrounding}{The surrounding nucleotides with a . at the mutation site}
}}
\usage{
testDataSet
}
\description{
A random generated mutation dataset
}
\keyword{datasets}
